<feat> Add custom-head reader model
