+++ reader/__init__.py
@@ -1,4 +1,2 @@
 from .base_reader import BaseReader, DprReader
-from .lstm_reader import LstmHeadReader
-from .cnn_reader import CnnHeadReader
-from .fc_reader import FcHeadReader
+from .custom_reader import CustomHeadReader
